Converted to ts and linted
