+++ src/model/Context.ts
@@ -1,5 +1,3 @@
-
 /** The information available to the AI when calculating the scores. */
 interface Context {
-	
-} 
+}
