+++ src/model/Qualifier.ts
@@ -1,6 +1,6 @@
 /** Calculates a score that represents the utility/usefulness of its associated action. */
 interface Qualifier {
-	scorers: Scorer[];
-	action: Action;
-	score(context: Context) : Promise<number>;
+    scorers: Scorer[];
+    action: Action;
+    score(context: Context): Promise<number>;
 }
