+++ src/model/Selector.ts
@@ -1,0 +1,6 @@
+/**
+ * Selects the best Qualifier from the Qualifiers attached to the Selector
+ */
+interface Selector {
+    select(context: Context, qualifiers: Qualifier[], defaultValue: number): Promise<Qualifier>;
+}
